refactor: relocate ProductExportOptions to Catalog namespace
