+++ backend/src/Anela.Heblo.API/Extensions/ServiceCollectionExtensions.cs
@@ -25,7 +25,7 @@ using Anela.Heblo.API.PDFPrints;
 using Anela.Heblo.Application.Features.BackgroundJobs.Services;
 using Anela.Heblo.Adapters.FileSystem;
 using Anela.Heblo.API.Features.ExpeditionList;
-using Anela.Heblo.Application.Features.FileStorage;
+using Anela.Heblo.Application.Features.Catalog.Infrastructure;
 using Anela.Heblo.Application.Shared.Printing;
 using Anela.Heblo.Application.Features.Manufacture.UseCases.GetManufactureProtocol;
 using Anela.Heblo.Application.Features.Manufacture.UseCases.GetSemiproductRecipePdf;
